requirements.txt
# Core dependencies
streamlit==1.29.0
openai==1.6.1
chromadb==0.4.22
pypdf2==3.0.1
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==2.2.2
tiktoken==0.5.2
python-dotenv==1.0.0
python-docx==1.1.0
flashrank==0.2.0

# Additional utilities
numpy==1.24.3
pandas==2.0.3